Refactor: extract common EncodedSequenceRow
